backend/requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
supabase
PyJWT==2.8.0
flask-cors==4.0.0
yfinance==0.2.37